+++ backend/src/neuroagent/tools/context_analyzer_tool.py
@@ -20,7 +20,7 @@ class ContextAnalyzerInput(BaseModel):
 class ContextAnalyzerMetdata(EntitycoreMetadata):
     """Metadata for the Context Analyzer tool."""
 
-    frontend_url: str
+    current_frontend_url: str
 
 
 class ContextAnalyzerOutput(BaseModel):
@@ -53,10 +53,10 @@ class ContextAnalyzerTool(BaseTool):
         -------
             Description of the current page the user is on, formatted as a string.
         """
-        if not self.metadata.frontend_url:
+        if not self.metadata.current_frontend_url:
             raise ValueError("Please provide the current frontend url.")
 
-        url = self.metadata.frontend_url
+        url = self.metadata.current_frontend_url
 
         # retreive description
         with (Path(__file__).parent.parent / "platform_description.json").open() as f: