+++ backend/src/neuroagent/tools/entitycore_electricalcellrecording_getone.py
@@ -81,7 +81,15 @@ class ElectricalCellRecordingGetOneTool(BaseTool):
             raise ValueError(
                 f"The electrical cell recording endpoint returned a non 200 response code. Error: {response.text}"
             )
-        return ElectricalCellRecordingRead(**response.json())
+        # Add the link
+        response_json = response.json()
+        response_json["url_link"] = (
+            self.metadata.entity_frontend_url
+            + "/explore/interactive/experimental"
+            + "/electrophysiology/"
+            + response_json["id"]
+        )
+        return ElectricalCellRecordingRead(**response_json)
 
     @classmethod
     async def is_online(cls, *, httpx_client: AsyncClient, entitycore_url: str) -> bool: